parser/requirements.txt
selenium
requests
beautifulsoup4
mysql-connector-python